+++ src/main/java/net/frozenblock/lib/FrozenClient.java
@@ -34,7 +34,6 @@ import net.frozenblock.lib.screenshake.api.client.ScreenShaker;
 import net.frozenblock.lib.sound.api.FlyBySoundHub;
 import net.frozenblock.lib.sound.impl.block_sound_group.BlockSoundGroupManager;
 import net.frozenblock.lib.wind.api.ClientWindManager;
-import net.frozenblock.lib.wind.api.WindDisturbances;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.PackType;
@@ -73,9 +72,11 @@ public final class FrozenClient implements ClientModInitializer {
 		ClientTickEvents.START_WORLD_TICK.register(ClientWindManager::tick);
 		ClientTickEvents.START_WORLD_TICK.register(ScreenShaker::tick);
 		ClientTickEvents.START_WORLD_TICK.register(level -> FlyBySoundHub.update(Minecraft.getInstance(), Minecraft.getInstance().getCameraEntity(), true));
-		ClientTickEvents.START_CLIENT_TICK.register(client -> WindDisturbances.clearAndSwitchClient());
-		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> ScreenShaker.clear());
-		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> WindDisturbances.clearAllClient());
+		ClientTickEvents.START_CLIENT_TICK.register(client -> ClientWindManager.clearAndSwitchWindDisturbances());
+		ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> {
+			ScreenShaker.clear();
+			ClientWindManager.clearAllWindDisturbances();
+		});
 	}
 
 }