+++ src/main/java/net/frozenblock/lib/wind/mixin/AbstractWindChargeMixin.java
@@ -18,46 +18,22 @@
 
 package net.frozenblock.lib.wind.mixin;
 
+import net.frozenblock.lib.wind.api.WindDisturbanceLogic;
 import net.frozenblock.lib.wind.api.WindDisturbingEntity;
-import net.frozenblock.lib.wind.impl.WindDisturbance;
-import net.minecraft.util.Mth;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.projectile.windcharge.AbstractWindCharge;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.AABB;
-import net.minecraft.world.phys.Vec3;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 
 @Mixin(AbstractWindCharge.class)
 public abstract class AbstractWindChargeMixin implements WindDisturbingEntity {
-	@Unique
-	private static final double FROZENLIB$WIND_RANGE_WIND_CHARGE = 5D;
 
 	@Unique
 	@Nullable
 	@Override
-	public WindDisturbance.DisturbanceLogic frozenLib$makeDisturbanceLogic() {
-		return this::frozenLib$calculateBreezeWindAndWeight;
-	}
-
-	@Unique
-	@Nullable
-	private WindDisturbance.DisturbanceResult frozenLib$calculateBreezeWindAndWeight(Level level, @NotNull Vec3 windOrigin, AABB affectedArea, Vec3 windTarget) {
-		double distance = windOrigin.distanceTo(windTarget);
-		if (distance <= FROZENLIB$WIND_RANGE_WIND_CHARGE) {
-			Vec3 chargeMovement = AbstractWindCharge.class.cast(this).getDeltaMovement();
-			double strengthFromDistance = Mth.clamp((FROZENLIB$WIND_RANGE_WIND_CHARGE - distance) / (FROZENLIB$WIND_RANGE_WIND_CHARGE * 0.5D), 0D, 1D);
-			Vec3 windVec = new Vec3(chargeMovement.x, chargeMovement.y, chargeMovement.z).scale(3D * strengthFromDistance);
-
-			return new WindDisturbance.DisturbanceResult(
-				strengthFromDistance,
-				(FROZENLIB$WIND_RANGE_WIND_CHARGE - distance) * 2D,
-				windVec
-			);
-		}
-		return null;
+	public ResourceLocation frozenLib$getWindDisturbanceLogicID() {
+		return WindDisturbanceLogic.WIND_CHARGE;
 	}
 
 	@Unique