+++ src/main/java/net/frozenblock/lib/wind/mixin/BreezeMixin.java
@@ -18,55 +18,22 @@
 
 package net.frozenblock.lib.wind.mixin;
 
-import net.frozenblock.lib.math.api.AdvancedMath;
+import net.frozenblock.lib.wind.api.WindDisturbanceLogic;
 import net.frozenblock.lib.wind.api.WindDisturbingEntity;
-import net.frozenblock.lib.wind.impl.WindDisturbance;
-import net.minecraft.util.Mth;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.monster.breeze.Breeze;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.AABB;
-import net.minecraft.world.phys.Vec3;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 
 @Mixin(Breeze.class)
 public abstract class BreezeMixin implements WindDisturbingEntity {
-	@Unique
-	private static final double FROZENLIB$WIND_RANGE_BREEZE = 6D;
 
 	@Unique
 	@Nullable
 	@Override
-	public WindDisturbance.DisturbanceLogic frozenLib$makeDisturbanceLogic() {
-		return this::frozenLib$calculateBreezeWindAndWeight;
-	}
-
-	@Unique
-	@Nullable
-	private WindDisturbance.DisturbanceResult frozenLib$calculateBreezeWindAndWeight(Level level, @NotNull Vec3 windOrigin, AABB affectedArea, Vec3 windTarget) {
-		double distance = windOrigin.distanceTo(windTarget);
-		if (distance <= FROZENLIB$WIND_RANGE_BREEZE) {
-			Vec3 breezeLookVec = Breeze.class.cast(this).getForward();
-			Vec3 differenceInPoses = windOrigin.subtract(windTarget);
-			double scaledDistance = (FROZENLIB$WIND_RANGE_BREEZE - distance) / FROZENLIB$WIND_RANGE_BREEZE;
-			double strengthFromDistance = Mth.clamp((FROZENLIB$WIND_RANGE_BREEZE - distance) / (FROZENLIB$WIND_RANGE_BREEZE * 0.75D), 0D, 1D);
-			double angleBetween = AdvancedMath.getAngleBetweenXZ(breezeLookVec, differenceInPoses);
-
-			double x = Math.cos((angleBetween * Math.PI) / 180D);
-			double z = -Math.sin((angleBetween * Math.PI) / 180D);
-			x = -Mth.lerp(scaledDistance, (x - (differenceInPoses.x * 0.45D)) * 0.5D, x);
-			z = -Mth.lerp(scaledDistance, (z - (differenceInPoses.z * 0.45D)) * 0.5D, z);
-
-			Vec3 windVec = new Vec3(x, strengthFromDistance, z).scale(1D);
-			return new WindDisturbance.DisturbanceResult(
-				strengthFromDistance,
-				FROZENLIB$WIND_RANGE_BREEZE - distance,
-				windVec
-			);
-		}
-		return null;
+	public ResourceLocation frozenLib$getWindDisturbanceLogicID() {
+		return WindDisturbanceLogic.BREEZE;
 	}
 
 	@Unique